removed util dir, changed log to logrus
