+++ server/Go/server-1.0/webclients/appclient_commands.txt
@@ -1,0 +1,30 @@
+// Websocket requests
+
+{"action":"get", "path":"Vehicle.Cabin.Door.*.*.IsOpen", "requestId":"123"}
+
+{"action":"get", "path":"Vehicle.Cabin.Door.Row1.Right.IsOpen", "requestId":"234"}
+
+{"action":"subscribe", "path":"Vehicle.Cabin.Door.Row1.Right.IsOpen", "requestId":"234"}  // "filters" key-value left out
+
+{"action":"unsubscribe", "subscriptionId":"789", "requestId":"234"}
+
+{"action":"set", "path":"Vehicle.Cabin.Door.Row1.Right.IsOpen", "value":"999", "requestId":"234"}
+
+{"action":"getmetadata", "path":"Vehicle.Cabin.Door.Row1.Right?$spec", "requestId":"234"}
+
+
+// HTTP requests
+
+Get command
+HTTP GET 
+URL=Vehicle/Cabin/Door/Row1/Right/IsOpen
+
+Set command
+HTTP POST 
+URL=Vehicle/Cabin/Door/Row1/Right/IsOpen
+Payload=123
+
+Getmetadata command
+HTTP GET 
+URL=Vehicle/Cabin/Door/Row1/Right?$spec
+
