--- server/Go/server-1.0/wsclient2.html
@@ -1,74 +1,0 @@
-<!-- 
-/**
-* (C) 2019 Geotab
-* (C) 2019 Volvo Cars
-*
-* All files and artifacts in the repository at https://github.com/MEAE-GOT/W3C_VehicleSignalInterfaceImpl
-* are licensed under the provisions of the license provided by the LICENSE file in this repository.
-*
-**/
- -->
-
-<!DOCTYPE html>
-<html>
-<body>
-
-<h2>Websocket client example</h2>
-
-<div>
-<input id="host-ip" type="text" placeholder="host ip"/>
-<button onclick="setHostIP()">Server IP</button>
-</div>
-<div>
-<input id="input" type="text" placeholder="JSON payload"/>
-<button onclick="send()">Send</button>
-<pre id="output"></pre>
-</div>
-
-<script>
-    var externHostIp = document.getElementById("host-ip");
-    var hostIP = "Host unknown";
-    var input = document.getElementById("input");
-    var output = document.getElementById("output");
-    var sessionOpen = false;
-    var socket;
-
-    try {
-        function send() {
-            if (hostIP == "Host unknown") {
-                output.innerHTML += "Host unknown. Please specify.\n";
-                input.value = "";
-                return;
-            };
-            if (sessionOpen == true) {
-                socket.send(input.value);
-                input.value = "";
-            }
-        }
-    } catch(err) {
-        output.innerHTML += "Send error. " + err.message + "\n";
-    }
-
-    try {
-        function setHostIP() {
-            hostIP = externHostIp.value;
-                socket = new WebSocket("ws://" + hostIP + ":8080");
-
-                socket.onopen = function () {
-                    output.innerHTML += "Status: Connected\n";
-                };
-
-                socket.onmessage = function (e) {
-                    output.innerHTML += "Server: " + e.data + "\n";
-                };
-                sessionOpen = true;
-        }
-    } catch(err) {
-        output.innerHTML += "Host IP error. " + err.message + "\n";
-    }
-
-</script>
-
-</body>
-</html>
-
